Start using the color_attachment_array as a history buffer
- still renders as usual, but now we're using the array as a time buffer
- fix animation breakage on the pause menu thing

+++ src/main.rs
@@ -102,7 +102,7 @@ fn run(opt: Options) {
                                 label: Some("Render Encoder"),
                             },
                         );
-                        app_state.render(&window, &mut encoder, &output, frame_index);
+                        app_state.render(&window, &mut encoder, &output, frame_index as usize);
                         app_state
                             .gpu
                             .queue
